requirements.txt
beautifulsoup4~=4.12.3
pyTelegramBotAPI
requests~=2.31.0
python-dotenv~=1.0.1